+++ THERMOSTAT/thermostat-led-display/list-utils.ino
@@ -1,5 +1,5 @@
 template<typename T>
-T ListUtils::getByIndex(std::list<T> l, int i) {
+T ListUtils::getByIndex(std::list<T> l, short i) {
   // TODO maybe we should use the same exc-like mechanism as in ds18b20-sensor.ino.
   if (i >= l.size()) return nullptr;
   auto it = l.begin();
